Activity to edit created

+++ app/src/main/java/com/example/gabriela/agenda/model/Classmate.java
@@ -8,7 +8,7 @@ public class Classmate implements Serializable{
     private String address;
     private String phone;
     private String site;
-    private Double stars;
+    private int stars;
 
 
     public Long getId() {
@@ -55,7 +55,7 @@ public class Classmate implements Serializable{
         return stars;
     }
 
-    public void setStars(Double stars) {
+    public void setStars(int stars) {
         this.stars = stars;
     }
